Cleanup: change pattern + do SIGINT before SIGKILL
Some tests with the same name as daemon could be killed. This changes
the pattern so that it does not happen.

Moreover, it is safer for some daemon to receive a SIGINT to clean
whatever they want before resorting to a hard SIGKILL after a few
seconds if that did not work.